Support for larger file size

+++ src/main/java/se/anyro/tgbotapi/types/file/Video.java
@@ -10,5 +10,5 @@ public class Video {
     public int duration;
     public PhotoSize thumb;
     public String mime_type;
-    public int file_size;
+    public long file_size;
 }
